fix: 主干告警清零 【背景】 主干告警清零
【修改方案】
在有符号数与无符号数混用的地方添加了强制类型转换

【影响】
对现有的产品编译不会有影响。

re #I4SBQE

+++ components/net/lwip-2.1/porting/src/sys_arch.c
@@ -67,7 +67,7 @@ sys_thread_t sys_thread_new(const char *name, lwip_thread_fn thread, void *arg, 
     task.pfnTaskEntry = (TSK_ENTRY_FUNC)thread;
     task.uwStackSize = stackSize;
     task.pcName = (char *)name;
-    task.usTaskPrio = prio;
+    task.usTaskPrio = (UINT16)prio;
     task.uwArg = (UINTPTR)arg;
     task.uwResved = LOS_TASK_STATUS_DETACHED;
     ret = LOS_TaskCreate(&taskID, &task);
